style: add footer styles and adaptive and fix styles errors
