+++ .prettier.config.cjs
@@ -2,8 +2,9 @@ module.exports = {
     semi: true,
     singleQuote: false,
     trailingComma: 'all',
-    printWidth: 100,
+    printWidth: 120,
     tabWidth: 2,
     bracketSpacing: true,
+    jsxBracketSameLine: false,
     arrowParens: 'always',
 };
