+++ extractedTranslations/en/translation.json
@@ -150,6 +150,9 @@
   "Любые материалы, файлы и сервисы, содержащиеся на сайте, не могут быть воспроизведены в какой-либо форме, каким-либо способом, полностью или частично без предварительного письменного\nразрешения Компании, за исключением случаев, указанных в Соглашении об использовании сайта": {
     "": ""
   },
+  "Любые материалы, файлы и сервисы, содержащиеся на сайте, не могут быть воспроизведены полностью или частично без предварительного письменного разрешения компании, \n                        за исключением случаев, предусмотренных правилами использования сайта": {
+    "": ""
+  },
   "Любые материалы, файлы и сервисы, содержащиеся на сайте, не могут быть воспроизведены полностью или частично без предварительного письменного разрешения компании, за исключением случаев, предусмотренных правилами использования сайта": {
     "": ""
   },
