+++ src/widgets/Footer/ui/FooterContentColumns/FooterContentColumns.stories.tsx
@@ -1,0 +1,29 @@
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { Theme } from '@/shared/const/theme';
+import { StoreDecorator } from '@/shared/config/storybook/StoreDecorator/StoreDecorator';
+import { ThemeDecorator } from '@/shared/config/storybook/ThemeDecorator/ThemeDecorator';
+import { FooterContentColumns } from './FooterContentColumns';
+
+export default {
+    title: 'widget/FooterContentColumns',
+    component: FooterContentColumns,
+    argTypes: {
+        backgroundColor: { control: 'color' },
+    },
+} as ComponentMeta<typeof FooterContentColumns>;
+
+const Template: ComponentStory<typeof FooterContentColumns> = (args) => <FooterContentColumns {...args} />;
+
+export const Light = Template.bind({});
+Light.args = {};
+Light.decorators = [
+  ThemeDecorator(Theme.LIGHT),
+  StoreDecorator({ user: { authData: { /* заполняйте по необходимости */ } } })
+];
+
+export const Dark = Template.bind({});
+Dark.args = {};
+Dark.decorators = [
+  ThemeDecorator(Theme.DARK),
+  StoreDecorator({ user: { authData: { /* заполняйте по необходимости */ } } })
+];
